+++ packages/brick-kit/src/core/Runtime.ts
@@ -11,6 +11,7 @@ import {
   PluginRuntimeContext,
   BrickConf,
   TemplatePackage,
+  SiteMapItem,
 } from "@next-core/brick-types";
 import {
   Kernel,
@@ -119,6 +120,12 @@ export class Runtime implements AbstractRuntime {
     return apps;
   }
 
+  hasInstalledApp(appId: string): boolean {
+    return kernel.bootstrapData.microApps.some(
+      (app) => app.id === appId && app.installStatus !== "running"
+    );
+  }
+
   reloadMicroApps(interceptorParams?: InterceptorParams): Promise<void> {
     return kernel.loadMicroApps(
       {
@@ -136,6 +143,11 @@ export class Runtime implements AbstractRuntime {
   /* istanbul ignore next */
   getDesktops(): DesktopData[] {
     return kernel.bootstrapData.desktops || [];
+  }
+
+  /* istanbul ignore next */
+  getLaunchpadSiteMap(): SiteMapItem[] {
+    return kernel.bootstrapData.siteSort || [];
   }
 
   /* istanbul ignore next */