+++ packages/brick-scripts/template/libs-pkg/scripts/post-build.js
@@ -1,0 +1,3 @@
+const { postBuild } = require("@next-core/build-config-factory");
+
+postBuild("libs");