+++ packages/dev-dependencies/patches/index.js
@@ -4,3 +4,6 @@ exports.updateMRTemplates = require("./updateMRTemplates");
 exports.updateBuildStories = require("./updateBuildStories");
 exports.updateRenovateFileFilters = require("./updateRenovateFileFilters");
 exports.updateLicense = require("./updateLicense");
+exports.addPostBuildScriptForLibs = require("./addPostBuildScriptForLibs");
+exports.disableSdkRenovate = require("./disableSdkRenovate");
+exports.migrateOfficialRenovate = require("./migrateOfficialRenovate");