+++ packages/sdk-scripts/src/lib/Context.ts
@@ -2,10 +2,8 @@ import os from "os";
 import path from "path";
 import * as changeCase from "change-case";
 import prettier from "prettier";
-import yaml from "js-yaml";
 import { Api, Model } from "./internal";
 import { FileWithContent } from "../interface";
-import { normalizeSemver } from "../utils";
 
 export class Context {
   readonly apiMap = new Map<string, Api>();
@@ -21,7 +19,6 @@ export class Context {
   }
 
   toFiles(sdkRoot: string): FileWithContent[] {
-    const apiSuffix = "Api";
     const apiByModelIndexes = Array.from(
       this.apiByModelExportsMap.entries()
     ).map(([modelSeg, apis]) => ({
@@ -37,46 +34,24 @@ export class Context {
         models.map((model) => [`export * from "./${model}"`]).join(";"),
     }));
 
-    const sdkIndexImports = this.indexApiExports
+    const sdkIndexExports = this.indexApiExports
       .map(
-        (modelSeg) =>
-          `import * as ${changeCase.pascalCase(
-            modelSeg
-          )}${apiSuffix} from "./api/${this.serviceSeg}/${modelSeg}";`
+        (modelSeg) => `export * from "./api/${this.serviceSeg}/${modelSeg}";`
       )
       .concat(
         this.indexModelExports.map(
           (serviceSeg) =>
-            `import * as ${changeCase.pascalCase(
+            `export * as ${changeCase.pascalCase(
               serviceSeg
             )}Models from "./model/${serviceSeg}";`
         )
       )
       .join(os.EOL);
 
-    const sdkIndexExports = this.indexApiExports
-      .map((modelSeg) => changeCase.pascalCase(modelSeg) + apiSuffix)
-      .concat(
-        this.indexModelExports.map(
-          (serviceSeg) => `${changeCase.pascalCase(serviceSeg)}Models`
-        )
-      )
-      .join(",");
     const sdkIndex = {
       filePath: "./index",
-      toString: () =>
-        `${sdkIndexImports}${os.EOL}export { ${sdkIndexExports} }`,
+      toString: () => sdkIndexExports,
     };
-
-    const contractsYaml: FileWithContent = [
-      path.join(sdkRoot, "deploy", "contracts.yaml"),
-      yaml.safeDump({
-        depend_contracts: Array.from(this.apiMap.values()).map((api) => ({
-          name: `easyops.api.${this.serviceSeg}.${api.modelSeg}.${api.originalName}`,
-          version: `^${normalizeSemver(api.doc.version)}`,
-        })),
-      }),
-    ];
 
     return [
       ...Array.from(this.apiMap.values()),
@@ -84,11 +59,9 @@ export class Context {
       ...apiByModelIndexes,
       ...modelByServiceIndexes,
       sdkIndex,
-    ]
-      .map<FileWithContent>((file) => [
-        path.join(sdkRoot, "src", file.filePath + ".ts"),
-        prettier.format(file.toString(), { parser: "typescript" }),
-      ])
-      .concat([contractsYaml]);
+    ].map<FileWithContent>((file) => [
+      path.join(sdkRoot, "src", file.filePath + ".ts"),
+      prettier.format(file.toString(), { parser: "typescript" }),
+    ]);
   }
 }