+++ packages/sdk-scripts/src/lib/FunctionBlock.ts
@@ -34,7 +34,14 @@ export class FunctionBlock {
   private getApiUrlAndParams(): { url: string; args: string[] } {
     const uri = this.api.doc.endpoint.uri.replace(/^\//, "");
     const uriParams = getParamsInUri(uri);
-    let url = `${FunctionBlock.uriPrefix}/${this.api.serviceName}/`;
+
+    // If the sdk is for api_gateway itself,
+    // the url prefix and service name should be ignored.
+    let url =
+      this.api.context.serviceSeg === "api_gateway"
+        ? ""
+        : `${FunctionBlock.uriPrefix}/${this.api.serviceName}/`;
+
     let quote: string;
     if (uriParams.length > 0) {
       quote = "`";
@@ -101,8 +108,12 @@ export class FunctionBlock {
       returnBlock = `http.${api.method.realName}<${responseBodyTypeName}>(${url},${callArgsString});`;
     }
 
+    const contractComment = `/**! @contract easyops.api.${this.api.serviceName} */`;
+
     if (isFormData) {
-      returnBlock = `{ ${FunctionBlock.formDataBlock} return ${returnBlock}; }`;
+      returnBlock = `{ ${contractComment} ${FunctionBlock.formDataBlock} return ${returnBlock}; }`;
+    } else {
+      returnBlock = `${contractComment} ${returnBlock}`;
     }
 
     const annotations = [];
@@ -119,7 +130,7 @@ export class FunctionBlock {
     } */`;
 
     return `${content}${os.EOL}export const ${
-      api.displayName
+      api.exportName
     } = ${asyncPrefix}(${args.join(
       ","
     )}): Promise<${responseBodyTypeName}> => ${returnBlock}`;